+++ packages/jupyterlab-drawio-pdf/src/tokens.ts
@@ -1,0 +1,5 @@
+import { CMD_NS } from '@deathbeds/jupyterlab-drawio/lib/tokens';
+
+export namespace CommandIds {
+  export const provision = `${CMD_NS}:provision`;
+}