+++ py_packages/jupyter-drawio-export/src/jupyter_drawio_export/constants.py
@@ -1,0 +1,5 @@
+PNG_DRAWIO_INFO = "mxfile"
+DRAWIO_APP = (
+    "static/node_modules/@deathbeds/jupyterlab-drawio-webpack/drawio/src/main/webapp"
+)
+IPYNB_METADATA = "@deathbeds/jupyterlab-drawio"