+++ py_packages/jupyter-drawio-export/src/jupyter_drawio_export/tests/test_provision.py
@@ -1,0 +1,3 @@
+def test_provision(provision_app, tmp_path):
+    provision_app.start()
+    assert [*tmp_path.glob("*")]